backend/requirements.txt
flask
numpy
opencv-python
python-dotenv
